requirements.txt
streamlit
pillow
numpy
gtts
pyttsx3
